==> 01_MimicIV/02_ICU_stays.ipynb <==
## ICU Stays

Export file -> CSV\Exports\o02_ICU_Stays.csv

Export file -> CSV\Exports\o03_ICU_First_Stays.csv

-------------------------------

In this part of the code I get as return the ICU stays of the patients diagnosed with the disease I have declared using the ICDs in the csv file.

I will keep for each patient only their first admission in the ICU. I will sort the results based on subject_id, hadm_id and intime and I will keep only the first record for each new subject_id I find by scanning the array.

import gc
import gzip
import pandas as pd

# Compressed file paths
compressed_file_path = r"..\\00_Datasets\\mimic-iv-3_1\\icu\\icustays.csv.gz"

# Read the stroke_diagnosis.csv file
diagnosis_df = pd.read_csv(r"CSV/Exports/o01_diagnosis.csv")

# Read and pass the compressed CSV file into a DataFrame
icustays_df = pd.read_csv(compressed_file_path, compression='gzip')

# Merge the two DataFrames on the hadm_id and subject_id columns
merged_df = pd.merge(icustays_df, diagnosis_df, on=["subject_id", "hadm_id"])

# Export the merged DataFrame to a CSV file
merged_df.to_csv(r'CSV/Exports/o02_icu_stay.csv', index=False)

# Sort by 'subject_id', 'hadm_id', 'intime'
merged_df.sort_values(by=['subject_id', 'hadm_id', 'intime'], inplace=True)

# Get the unique subject_id
unique_subject_ids = merged_df['subject_id'].unique()

# Keep only the first row of every unique subject_id
first_rows_df = merged_df[merged_df['subject_id'].isin(unique_subject_ids)].groupby('subject_id').first().reset_index()

# Export the first rows DataFrame to a CSV file
first_rows_df.to_csv(r'CSV/Exports/o03_icu_first_stay.csv', index=False)

# Free RAM
gc.collect()
