==> 01_MimicIV/05_ICU_group_mean_etc.ipynb <==
## Chartevent Groupby 2nd Phase
Export the scrolling observation window (Mean, Median, Min, Max) sorted to unique separate files.

import pandas as pd
import os
import gc

# Directory where the CSV files are located
csv_directory = 'CSV/Exports/groupby/chartevent/mean/'

# Initialize an empty list to collect DataFrames
dfs = []

# Iterate through all CSV files in the directory
for filename in os.listdir(csv_directory):
    if filename.endswith('.csv'):
        file_path = os.path.join(csv_directory, filename)
        # Read the CSV file
        df = pd.read_csv(file_path)
        # Filter rows with 'Count' equal to 1 and append to the list
        filtered_rows = df[df['Count'] == 1]
        dfs.append(filtered_rows)

# Concatenate all DataFrames in the list into a single DataFrame
combined_data = pd.concat(dfs, ignore_index=True)

# Sort
combined_data.sort_values(by=['subject_id','itemid','Time_Zone'], inplace=True)

# Save to a CSV file
combined_data.to_csv(r'CSV\Exports\groupby\chartevent\o01_icu_chartevent_grouped_mean.csv', index=False)

# Directory where the CSV files are located
csv_directory = 'CSV/Exports/groupby/chartevent/median/'

# Initialize an empty list to collect DataFrames
dfs = []

# Iterate through all CSV files in the directory
for filename in os.listdir(csv_directory):
    if filename.endswith('.csv'):
        file_path = os.path.join(csv_directory, filename)
        # Read the CSV file
        df = pd.read_csv(file_path)
        # Filter rows with 'Count' equal to 1 and append to the list
        filtered_rows = df[df['Count'] == 1]
        dfs.append(filtered_rows)

# Concatenate all DataFrames in the list into a single DataFrame
combined_data = pd.concat(dfs, ignore_index=True)

# Sort
combined_data.sort_values(by=['subject_id','itemid','Time_Zone'], inplace=True)

# Save to a CSV file
combined_data.to_csv(r'CSV\Exports\groupby\chartevent\o02_icu_chartevent_grouped_median.csv', index=False)

# Directory where the CSV files are located
csv_directory = 'CSV/Exports/groupby/chartevent/min/'

# Initialize an empty list to collect DataFrames
dfs = []

# Iterate through all CSV files in the directory
for filename in os.listdir(csv_directory):
    if filename.endswith('.csv'):
        file_path = os.path.join(csv_directory, filename)
        # Read the CSV file
        df = pd.read_csv(file_path)
        # Filter rows with 'Count' equal to 1 and append to the list
        filtered_rows = df[df['Count'] == 1]
        dfs.append(filtered_rows)

# Concatenate all DataFrames in the list into a single DataFrame
combined_data = pd.concat(dfs, ignore_index=True)

# Sort
combined_data.sort_values(by=['subject_id','itemid','Time_Zone'], inplace=True)

# Save to a CSV file
combined_data.to_csv(r'CSV\Exports\groupby\chartevent\o03_icu_chartevent_grouped_min.csv', index=False)

# Directory where the CSV files are located
csv_directory = 'CSV/Exports/groupby/chartevent/max/'

# Initialize an empty list to collect DataFrames
dfs = []

# Iterate through all CSV files in the directory
for filename in os.listdir(csv_directory):
    if filename.endswith('.csv'):
        file_path = os.path.join(csv_directory, filename)
        # Read the CSV file
        df = pd.read_csv(file_path)
        # Filter rows with 'Count' equal to 1 and append to the list
        filtered_rows = df[df['Count'] == 1]
        dfs.append(filtered_rows)

# Concatenate all DataFrames in the list into a single DataFrame
combined_data = pd.concat(dfs, ignore_index=True)

# Sort
combined_data.sort_values(by=['subject_id','itemid','Time_Zone'], inplace=True)

# Save to a CSV file
combined_data.to_csv(r'CSV\Exports\groupby\chartevent\o04_icu_chartevent_grouped_max.csv', index=False)

# Free RAM
gc.collect()
