==> 01_MimicIV/08_produce_datasets.ipynb <==
import gzip
import pandas as pd

# Read the exported for header chartevent CSV file
mean_df = pd.read_csv(r'CSV\Exports\datasets\Temp\o01_mean_table.csv')
median_df = pd.read_csv(r'CSV\Exports\datasets\Temp\o02_median_table.csv')
min_df = pd.read_csv(r'CSV\Exports\datasets\Temp\o03_min_table.csv')
max_df = pd.read_csv(r'CSV\Exports\datasets\Temp\o04_max_table.csv')
demographics_df = pd.read_csv(r'CSV\Exports\o08_demographics.csv')

# Merge mean with demographics
temp_df = pd.merge(mean_df, demographics_df, on=['subject_id', 'hadm_id'])

# Set new order
new_order = ['subject_id', 'hadm_id', 'Time_Zone', 'gender', 'age', 'language', 'marital_status', 'race'
            ] + [col for col in temp_df.columns if col not in 
                 ['subject_id', 'hadm_id', 'Time_Zone', 'gender', 'age', 'language', 'marital_status', 'race']]

# Reorder the columns in the DataFrame
temp_df = temp_df[new_order]

# Add a new column 'row_count' at the beginning
# I'm going to use it as a split point when I load
# the dataset in the rapidminer with filter operator
temp_df.insert(0, 'row_count', range(1, len(temp_df) + 1))

# Reset the index to match the row_count
temp_df.reset_index(drop=True, inplace=True)

# Replace from hospital_expire_flag 0 with survive and 1 with death
temp_df['hospital_expire_flag'] = temp_df['hospital_expire_flag'].replace({0: 'Survive', 1: 'Death'})

# Export the merged DataFrame to a CSV file
temp_df.to_csv(r'CSV\Exports\datasets\whole_set\o01_final_mean_table.csv', index=False)

# Delete dataframes to free memory.
del (temp_df)

# Merge median with demographics
temp_df = pd.merge(median_df, demographics_df, on=['subject_id', 'hadm_id'])

# Set new order
new_order = ['subject_id', 'hadm_id', 'Time_Zone', 'gender', 'age', 'language', 'marital_status', 'race'
            ] + [col for col in temp_df.columns if col not in 
                 ['subject_id', 'hadm_id', 'Time_Zone', 'gender', 'age', 'language', 'marital_status', 'race']]

# Reorder the columns in the DataFrame
temp_df = temp_df[new_order]

# Add a new column 'row_count' at the beginning
# I'm going to use it as a split point when I load
# the dataset in the rapidminer with filter operator
temp_df.insert(0, 'row_count', range(1, len(temp_df) + 1))

# Reset the index to match the row_count
temp_df.reset_index(drop=True, inplace=True)

# Replace from hospital_expire_flag 0 with survive and 1 with death
temp_df['hospital_expire_flag'] = temp_df['hospital_expire_flag'].replace({0: 'Survive', 1: 'Death'})

# Export the merged DataFrame to a CSV file
temp_df.to_csv(r'CSV\Exports\datasets\whole_set\o02_final_median_table.csv', index=False)

# Delete dataframes to free memory.
del (temp_df)

# Merge min with demographics
temp_df = pd.merge(min_df, demographics_df, on=['subject_id', 'hadm_id'])

# Set new order
new_order = ['subject_id', 'hadm_id', 'Time_Zone', 'gender', 'age', 'language', 'marital_status', 'race'
            ] + [col for col in temp_df.columns if col not in 
                 ['subject_id', 'hadm_id', 'Time_Zone', 'gender', 'age', 'language', 'marital_status', 'race']]

# Reorder the columns in the DataFrame
temp_df = temp_df[new_order]

# Add a new column 'row_count' at the beginning
# I'm going to use it as a split point when I load
# the dataset in the rapidminer with filter operator
temp_df.insert(0, 'row_count', range(1, len(temp_df) + 1))

# Reset the index to match the row_count
temp_df.reset_index(drop=True, inplace=True)

# Replace from hospital_expire_flag 0 with survive and 1 with death
temp_df['hospital_expire_flag'] = temp_df['hospital_expire_flag'].replace({0: 'Survive', 1: 'Death'})

# Export the merged DataFrame to a CSV file
temp_df.to_csv(r'CSV\Exports\datasets\whole_set\o03_final_min_table.csv', index=False)

# Delete dataframes to free memory.
del (temp_df)

# Merge max with demographics
temp_df = pd.merge(max_df, demographics_df, on=['subject_id', 'hadm_id'])

# Set new order
new_order = ['subject_id', 'hadm_id', 'Time_Zone', 'gender', 'age', 'language', 'marital_status', 'race'
            ] + [col for col in temp_df.columns if col not in 
                 ['subject_id', 'hadm_id', 'Time_Zone', 'gender', 'age', 'language', 'marital_status', 'race']]

# Reorder the columns in the DataFrame
temp_df = temp_df[new_order]

# Add a new column 'row_count' at the beginning
# I'm going to use it as a split point when I load
# the dataset in the rapidminer with filter operator
temp_df.insert(0, 'row_count', range(1, len(temp_df) + 1))

# Reset the index to match the row_count
temp_df.reset_index(drop=True, inplace=True)

# Replace from hospital_expire_flag 0 with survive and 1 with death
temp_df['hospital_expire_flag'] = temp_df['hospital_expire_flag'].replace({0: 'Survive', 1: 'Death'})

# Export the merged DataFrame to a CSV file
temp_df.to_csv(r'CSV\Exports\datasets\whole_set\o04_final_max_table.csv', index=False)

# Delete dataframes to free memory.
del (temp_df)
