==> 01_MimicIV/09_inlcude_los.ipynb <==
import pandas as pd

# Load the two CSV files
df_los = pd.read_csv(r'CSV\Exports\o03_icu_first_stay.csv')
df_final = pd.read_csv(r'CSV\Exports\datasets\whole_set\o01_final_mean_table.csv')

# Merge the dataframes on 'subject_id' and 'hadm_id'
merged_df = pd.merge(df_final, df_los[['subject_id', 'hadm_id', 'los']], on=['subject_id', 'hadm_id'], how='left')

# Save the result to a new CSV file
merged_df.to_csv(r'CSV\Exports\datasets\whole_set\o01_final_mean_with_los.csv', index=False)

#---------------------------------

df_final = pd.read_csv(r'CSV\Exports\datasets\whole_set\o02_final_median_table.csv')

# Merge the dataframes on 'subject_id' and 'hadm_id'
merged_df = pd.merge(df_final, df_los[['subject_id', 'hadm_id', 'los']], on=['subject_id', 'hadm_id'], how='left')

# Save the result to a new CSV file
merged_df.to_csv(r'CSV\Exports\datasets\whole_set\o02_final_median_with_los.csv', index=False)

#---------------------------------

df_final = pd.read_csv(r'CSV\Exports\datasets\whole_set\o03_final_min_table.csv')

# Merge the dataframes on 'subject_id' and 'hadm_id'
merged_df = pd.merge(df_final, df_los[['subject_id', 'hadm_id', 'los']], on=['subject_id', 'hadm_id'], how='left')

# Save the result to a new CSV file
merged_df.to_csv(r'CSV\Exports\datasets\whole_set\o03_final_min_with_los.csv', index=False)

#---------------------------------

df_final = pd.read_csv(r'CSV\Exports\datasets\whole_set\o04_final_max_table.csv')

# Merge the dataframes on 'subject_id' and 'hadm_id'
merged_df = pd.merge(df_final, df_los[['subject_id', 'hadm_id', 'los']], on=['subject_id', 'hadm_id'], how='left')

# Save the result to a new CSV file
merged_df.to_csv(r'CSV\Exports\datasets\whole_set\o04_final_max_with_los.csv', index=False)
